scripts/command/pyproject.toml
[project]
name = "command"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dynaconf>=3.2.12",
    "gitpython>=3.1.45",
    "jinja2>=3.1.6",
    "kubernetes>=34.1.0",
    "pydantic>=2.12.4",
    "python-on-whales>=0.79.0",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "rich>=14.2.0",
    "ruff>=0.14.7",
    "sqlalchemy>=2.0.44",
    "toml>=0.10.2",
    "typer>=0.20.0",
    "uv>=0.9.10",
]

[project.optional-dependencies]
nuitka-build = ["nuitka>=2.1.0", "zstandard>=0.22.0"]

[tool.ruff.lint]
select = [
    "E",  # https://docs.astral.sh/ruff/rules/#error-e
    "F",  # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "I",  # https://docs.astral.sh/ruff/rules/#isort-i
    "UP", # https://docs.astral.sh/ruff/rules/#pyupgrade-up
]
ignore = ["F401", "E501"]

[tool.pyright]
exclude = [ #
    ".git",
    "**/__pycache__",
    ".cache",
    ".idea",
    ".vscode",
    ".venv",
    ".devbox",
    "db/",
    "data/",
    "temp/",
    "logs/",
    "vendor/",
    "output/",
]

[dependency-groups]
dev = []
